+++ files/file.go
@@ -1,0 +1,10 @@
+package files
+
+import "io"
+
+type File interface {
+	io.Reader
+	io.Writer
+	io.Seeker
+	io.Closer
+}